models and DB structure fixes

+++ src/main/resources/db/changeLog/v.0.1.1-init/40-create-product-table.sql
@@ -1,14 +1,14 @@
 CREATE TABLE IF NOT EXISTS `product`
 (
-    `id`            int         NOT NULL AUTO_INCREMENT,
-    `manager_id`    int         NOT NULL,
-    `name`          varchar(70) NOT NULL,
-    `status`        int         NOT NULL,
-    `currency_code` int         NOT NULL,
-    `interest_rate` int         NOT NULL,
-    `limit`         int         NOT NULL,
-    `created_at`    timestamp   NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    `updated_at`    timestamp   NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    `id`            int           NOT NULL AUTO_INCREMENT,
+    `manager_id`    int           NOT NULL,
+    `name`          varchar(70)   NOT NULL,
+    `status`        int           NOT NULL,
+    `currency_code` int           NOT NULL,
+    `interest_rate` decimal(6, 4) NOT NULL,
+    `limit`         int           NOT NULL,
+    `created_at`    timestamp     NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `updated_at`    timestamp     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8;
